refactor: extract no-android logic into client package
