+++ android/library/src/main/java/io/logto/client/constant/CodeChallengeMethod.kt
@@ -1,4 +1,4 @@
-package io.logto.android.constant
+package io.logto.client.constant
 
 object CodeChallengeMethod {
     const val S256 = "S256"
