+++ android/library/src/main/java/io/logto/client/constant/GrantType.kt
@@ -1,4 +1,4 @@
-package io.logto.android.constant
+package io.logto.client.constant
 
 object GrantType {
     const val AUTHORIZATION_CODE = "authorization_code"
