+++ android/library/src/main/java/io/logto/client/constant/PromptValue.kt
@@ -1,4 +1,4 @@
-package io.logto.android.constant
+package io.logto.client.constant
 
 object PromptValue {
     const val CONSENT = "consent"
