+++ android/library/src/main/java/io/logto/client/constant/ResourceValue.kt
@@ -1,4 +1,4 @@
-package io.logto.android.constant
+package io.logto.client.constant
 
 object ResourceValue {
     const val LOGTO_API = "https://api.logto.io"
