+++ android/library/src/main/java/io/logto/client/constant/ResponseType.kt
@@ -1,4 +1,4 @@
-package io.logto.android.constant
+package io.logto.client.constant
 
 object ResponseType {
     const val CODE = "code"
