+++ android/library/src/main/java/io/logto/client/constant/StorageKey.kt
@@ -1,4 +1,4 @@
-package io.logto.android.constant
+package io.logto.client.constant
 
 object StorageKey {
     const val TOKEN_SET = "io.logto.tokenset"
