+++ android/library/src/main/java/io/logto/client/model/OidcConfiguration.kt
@@ -1,4 +1,4 @@
-package io.logto.android.model
+package io.logto.client.model
 
 data class OidcConfiguration(
     val authorizationEndpoint: String,
