+++ android/library/src/main/java/io/logto/client/utils/PkceUtils.kt
@@ -1,18 +1,16 @@
-package io.logto.android.pkce
+package io.logto.client.utils
 
-import android.util.Base64
 import com.aventrix.jnanoid.jnanoid.NanoIdUtils
-import io.logto.android.exception.LogtoException
+import io.logto.client.exception.LogtoException
+import org.jose4j.base64url.Base64Url
 import java.io.UnsupportedEncodingException
 import java.security.MessageDigest
 import java.security.NoSuchAlgorithmException
 
-object Pkce {
-
+object PkceUtils {
     private const val CODE_VERIFIER_ALPHABET =
         "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-._~"
     private const val CODE_VERIFIER_LEN = 64
-    private const val CODE_ENCODE_FLAG = Base64.URL_SAFE or Base64.NO_WRAP or Base64.NO_PADDING
     private const val DEFAULT_ALGORITHM = "SHA-256"
 
     fun generateCodeVerifier(): String {
@@ -21,15 +19,15 @@ object Pkce {
             CODE_VERIFIER_ALPHABET.toCharArray(),
             CODE_VERIFIER_LEN
         )
-        return Base64.encodeToString(randomString.toByteArray(), CODE_ENCODE_FLAG)
+        return Base64Url.encode(randomString.toByteArray())
     }
 
     fun generateCodeChallenge(codeVerifier: String): String {
         try {
             val digester: MessageDigest = MessageDigest.getInstance(DEFAULT_ALGORITHM)
-            digester.update(codeVerifier.toByteArray(Charsets.ISO_8859_1))
+            digester.update(codeVerifier.toByteArray(Charsets.UTF_8))
             val byteArray: ByteArray = digester.digest()
-            return Base64.encodeToString(byteArray, CODE_ENCODE_FLAG)
+            return Base64Url.encode(byteArray)
         } catch (exception: NoSuchAlgorithmException) {
             throw LogtoException(
                 "${LogtoException.ENCRYPT_ALGORITHM_NOT_SUPPORTED}: $DEFAULT_ALGORITHM",
