+++ android/library/src/test/java/io/logto/android/auth/browser/BrowserSignOutFlowTest.kt
@@ -4,13 +4,13 @@ import android.app.Activity
 import android.net.Uri
 import com.google.common.truth.Truth.assertThat
 import io.logto.android.client.LogtoApiClient
-import io.logto.android.config.LogtoConfig
-import io.logto.android.constant.QueryKey
-import io.logto.android.exception.LogtoException
-import io.logto.android.exception.LogtoException.Companion.EMPTY_REDIRECT_URI
-import io.logto.android.exception.LogtoException.Companion.INVALID_REDIRECT_URI
-import io.logto.android.exception.LogtoException.Companion.SIGN_OUT_FAILED
-import io.logto.android.model.OidcConfiguration
+import io.logto.client.config.LogtoConfig
+import io.logto.client.constant.QueryKey
+import io.logto.client.exception.LogtoException
+import io.logto.client.exception.LogtoException.Companion.EMPTY_REDIRECT_URI
+import io.logto.client.exception.LogtoException.Companion.INVALID_REDIRECT_URI
+import io.logto.client.exception.LogtoException.Companion.SIGN_OUT_FAILED
+import io.logto.client.model.OidcConfiguration
 import io.logto.android.utils.Utils
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
