+++ android/library/src/test/java/io/logto/android/storage/TokenSetStorageTest.kt
@@ -3,9 +3,9 @@ package io.logto.android.storage
 import android.app.Application
 import androidx.test.core.app.ApplicationProvider
 import com.google.common.truth.Truth.assertThat
-import io.logto.android.config.LogtoConfig
-import io.logto.android.constant.ScopeValue
-import io.logto.android.model.TokenSet
+import io.logto.client.config.LogtoConfig
+import io.logto.client.constant.ScopeValue
+import io.logto.client.model.TokenSet
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
